requirements.txt
Django~=3.2.7
djangorestframework~=3.12.4
gunicorn~=20.1.0
Pillow~=8.3.2